feat: implementando el MVC en los archivos
