+++ part_4/routes/movies.routes.js
@@ -1,86 +1,14 @@
 import { Router } from 'express'
-import { readJSON } from '../utils/createRequire.js'
-import { validateMovie, validatePartialMovie } from '../schemas/movies.js'
-import { randomUUID } from 'node:crypto'
-const movies = readJSON('../movies.json')
+import { MovieController } from '../controllers/movies.js'
 
 export const moviesRouter = Router()
 
-moviesRouter.get('/', (req, res) => {
-  // El header de origin no se envia cuando la peticion se hace al mismo origen
-  const { genre } = req.query
-  if (genre) {
-    const filteredMovies = movies.filter((movie) =>
-      movie.genre.some((g) => g.toLowerCase() === genre.toLowerCase())
-    )
-    return res.json(filteredMovies)
-  }
+moviesRouter.get('/', MovieController.getAll)
 
-  res.status(200).json(movies)
-})
+moviesRouter.get('/:id', MovieController.getBiId)
 
-moviesRouter.get('/:id', (req, res) => {
-  // path-to-regexp
-  const { id } = req.params
+moviesRouter.post('/', MovieController.create)
 
-  const movie = movies.find((movie) => movie.id === id)
-  if (movie) return res.status(200).json(movie)
+moviesRouter.delete('/:id', MovieController.delete)
 
-  res.status(404).json({ error: 'Movie not found' })
-})
-
-moviesRouter.post('/', (req, res) => {
-  const result = validateMovie(req.body) // utiliza zod para validar los datos del body
-
-  // devuelve un objeto con un array de errores
-  if (result.error) {
-    res.status(400).json({ error: result.error.errors })
-  }
-
-  // genera un nueva objeto con la data validada
-  const newMovie = {
-    id: randomUUID(),
-    ...result.data // los trespuntos es para desestructurar el objeto
-  }
-
-  movies.push(newMovie)// agrega la nueva pelicula al array de peliculas
-
-  res.status(201).json(newMovie)
-})
-
-moviesRouter.delete('/:id', (req, res) => {
-  const { id } = req.params
-  const movieIndex = movies.findIndex(movie => movie.id === id)
-
-  if (movieIndex === -1) {
-    return res.status(404).json({ message: 'Movie not found' })
-  }
-
-  movies.splice(movieIndex, 1)
-
-  return res.json({ message: 'Movie deleted' })
-})
-
-moviesRouter.patch('/:id', (req, res) => {
-  const { id } = req.params
-  const result = validatePartialMovie(req.body)
-
-  if (!result.success) {
-    return res.status(400).json({ error: JSON.parse(result.error.message) })
-  }
-
-  const movieIndex = movies.findIndex(movie => movie.id === id)
-
-  if (movieIndex === -1) {
-    return res.status(404).json({ message: 'Movie not found' })
-  }
-
-  const updateMovie = {
-    ...movies[movieIndex],
-    ...result.data
-  }
-
-  movies[movieIndex] = updateMovie
-
-  return res.status(200).json(updateMovie)
-})
+moviesRouter.patch('/:id', MovieController.update)
